# src/config/swagger.yaml
openapi: 3.0.3
info:
  title: SEEK API - Plateforme Immobilière
  version: 1.0.0
  description: |
    API pour la plateforme de location immobilère SEEK au Sénégal

    ## Environments
    - **Local**: http://localhost:8000
    - **Production**: https://seek-api.onrender.com

    ## Authentification
    L'API utilise plusieurs types d'authentification JWT:
    - **Admin**: Cookie HTTP-Only avec access/refresh tokens
    - **Owner**: Bearer token dans le header Authorization
    - **Locataire**: Bearer token spécifique
    - **Compte Public**: Bearer token spécifique
  contact:
    name: Support SEEK
    email: support@seek.sn
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:8000
    description: Serveur de développement local
  - url: https://seek-api.onrender.com
    description: Serveur de production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token pour Owner, Locataire, ou Compte Public
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: Cookie JWT pour Admin
...
paths:
...
  /api/owner/verification:
    post:
      tags:
        - Owner
      summary: Soumettre vérification identité
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vérification soumise
  /api/biens:
    get:
      tags:
        - Biems
      summary: Liste des biens du propriétaire
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste des biens
    post:
      tags:
        - Biems
      summary: Créer un bien (brouillon)
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Bien créé
  /api/biens/stats:
    get:
      tags:
        - Biems
      summary: Statistiques des biens du propriétaire
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Statistiques
...
  /api/types-logement/admin:
    get:
      tags:
        - Types
      summary: Tous les types (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des types
    post:
      tags:
        - Types
      summary: Créer un type de logement (admin)
      security:
        - cookieAuth: []
      responses:
        "201":
          description: Type créé
...
  /api/types-transaction/admin:
    get:
      tags:
        - Types
      summary: Tous les types (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des types
    post:
      tags:
        - Types
      summary: Créer un type de transaction (admin)
      security:
        - cookieAuth: []
      responses:
        "201":
          description: Type créé
...
  /api/owner/locataires:
    get:
      tags:
        - Locataire
      summary: Liste des locataires du propriétaire
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste des locataires
    post:
      tags:
        - Locataire
      summary: Créer un locataire
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Locataire créé
...
  /api/biens/{id}/bail:
    get:
      tags:
        - Baux
      summary: Bail actif d'un bien
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bail trouvé
    post:
      tags:
        - Baux
      summary: Créer un bail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Bail créé
...
  /api/public/favoris:
    get:
      tags:
        - Compte Public
      summary: Liste des favoris
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liste des favoris
...
  /api/transactions/admin:
    get:
      tags:
        - Transactions
      summary: Toutes les transactions (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Toutes les transactions
...
  /api/monetisation/admin/config:
    get:
      tags:
        - Monetisation
      summary: Configuration (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Configuration
    put:
      tags:
        - Monetisation
      summary: Modifier config (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Configuration modifiée
...
  /api/stats/admin:
    get:
      tags:
        - Stats
      summary: Statistiques admin
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Statistiques
  /api/stats/proprietaires:
    get:
      tags:
        - Stats
      summary: Statistiques des propriétaires
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Statistiques
...
  /api/owner:
    get:
      tags:
        - Owner
      summary: Score de confiance du propriétaire
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Score de confiance
  /api/annonces:
    get:
      tags:
        - Biems
      summary: Liste annonces (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des annonces
...
  /api/suspension:
    get:
      tags:
        - Auth
      summary: Liste suspensions (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des suspensions
...
  /api/modeles-contrat/admin:
    get:
      tags:
        - Baux
      summary: Tous les modèles (admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des modèles
    post:
      tags:
        - Baux
      summary: Créer un modèle (admin)
      security:
        - cookieAuth: []
      responses:
        "201":
          description: Modèle créé